+++ packages/core/src/index.ts
@@ -2,4 +2,3 @@ export { default, VOTJSError } from "./client";
 export { YandexSessionProtobuf, YandexVOTProtobuf } from "./protobuf";
 export * as VOTTypes from "./types/index";
 export * as videoData from "./utils/videoData";
-export * as votUtils from "./utils/vot";