+++ packages/core/src/types/index.ts
@@ -1,4 +1,3 @@
 export * as ClientType from "./client";
 export * as BaseHelperType from "./helpers/base";
-export * as VOTBackendType from "./vot";
 export * as YandexType from "./yandex";