+++ packages/ext/src/client.ts
@@ -1,10 +1,11 @@
 import VOTCoreClient from "@vot.js/core/client";
 import type { VOTOpts } from "@vot.js/core/types/client";
 import { browserSecHeaders } from "@vot.js/shared/secure";
-import type { VideoService } from "./types/service";
 import type { VOTProvider } from "@vot.js/core/types/providers/index";
 import type { YandexProvider } from "@vot.js/core/providers/yandex";
 import type { BaseProvider } from "@vot.js/core/providers/base";
+
+import type { VideoService } from "./types/service";
 
 export default class VOTClient<
   V extends string = VideoService,